skills/deal-pack/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "v23-deal-pack"
version = "0.1.0"
description = "Claude Code skill: normalize CRE DD materials into a deal pack."
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1",
    "pdfplumber>=0.11",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["deal_pack*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
